refactor: 예상치 못한 예외처리 시 error 메세지가 노출되지 않도록 변경

+++ src/main/java/gift/service/WishService.java
@@ -22,6 +22,9 @@ public class WishService {
     }
 
     public void update(WishPatchRequest request, Long memberId) {
+        if(!checkWishExist(request.productId(), memberId)) {
+            throw new EntityNotFoundException("Product with id " + request.productId() + " does not exist in wish");
+        }
         if (request.productCount() == 0) {
             deleteByProductId(request.productId(), memberId);
             return;
